fixdep: fix assumption that `long` is as large as `void *`
It is not. At least not in general. And in particular not on 64-bit
Windows, where sizeof(long) == 4.

What the code wanted to use is actually called intptr_t.

Note: `intptr_t` is a C99 feature.

+++ scripts/basic/fixdep.c
@@ -349,7 +349,7 @@ void do_config_file(char *filename)
 		return;
 	}
 	map = mmap(NULL, st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
-	if ((long) map == -1) {
+	if ((intptr_t) map == -1) {
 		perror("fixdep: mmap");
 		close(fd);
 		return;
@@ -424,7 +424,7 @@ void print_deps(void)
 		return;
 	}
 	map = mmap(NULL, st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
-	if ((long) map == -1) {
+	if ((intptr_t) map == -1) {
 		perror("fixdep: mmap");
 		close(fd);
 		return;
